+++ packages/flutter_blue_plus_android/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 4.0.5
+* Fixes to check for location services when plugin is called from a service
+
+## 4.0.3
+* Added option to disable Location Services check
+
+## 4.0.2
+* Fix compile error
+
+## 4.0.1
+* Fix compile error
+
+## 4.0.0
+* Use bytes instead of hex for platform communication (#1130)
+* Check if Android location services are enabled when doing scan (#1167)
+
 ## 3.0.0
 * Add option to provide pairing PIN to `createBond` (#1119)