+++ packages/flutter_blue_plus_darwin/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.0.1
+* fix unrecognized selector sent to instance (regression from 4.0.0)
+
+## 4.0.0
+* Use bytes instead of hex for platform communication (#1130)
+
 ## 3.0.0
 * Update platform interface version to 3.0.0